#7 added new idea submission, made report form as parent and widen it, removed commented part of code from githubApi
